refactor(auth): extract the persisted-auth cell codec into its own module
New persisted-auth-storage.ts owns the cell: the PersistedAuthStorage
type, the string<->cell codec (parsePersistedAuthCell, serializePersistedAuthCell),
and the synchronous Web Storage adapter. create-oauth-app-auth.ts imports
the type; machine-auth's write goes through serializePersistedAuthCell.

The codec earns extraction now that a third caller (the upcoming async
loader) shares the browser adapter's swallow-to-null read semantics.
machine-auth's read stays bespoke: it logs the discarded cell, the one
reader that does not swallow silently. Behavior-preserving.
