+++ packages/auth/src/node/machine-auth.ts
@@ -40,6 +40,7 @@ import {
 	type AuthFetch,
 	createOAuthAppAuth,
 } from '../create-oauth-app-auth.js';
+import { serializePersistedAuthCell } from '../persisted-auth-storage.js';
 import { createOobOAuthLauncher } from './oob-launcher.js';
 
 /**
@@ -199,7 +200,7 @@ async function saveMachineTokens(
 			}
 			await fs.mkdir(path.dirname(filePath), { recursive: true, mode: 0o700 });
 			const tmp = `${filePath}.tmp`;
-			await fs.writeFile(tmp, JSON.stringify(PersistedAuth.assert(value)), {
+			await fs.writeFile(tmp, serializePersistedAuthCell(value), {
 				mode: 0o600,
 			});
 			await fs.rename(tmp, filePath);
